Update Fiscal year

+++ database/migrations/2022_12_06_084744_create_tasks_table.php
@@ -20,11 +20,11 @@ class CreateTasksTable extends Migration
             $table->text('task_description')->nullable();
             $table->timestamp('task_start_date');
             $table->timestamp('task_end_date');
-            $table->integer('task_budget_gov_operating')->nullable();
-            $table->integer('task_budget_gov_investment')->nullable();
-            $table->integer('task_budget_gov_utility')->nullable();
-            $table->integer('task_budget_it_operating')->nullable();
-            $table->integer('task_budget_it_investment')->nullable();
+            $table->decimal('task_budget_gov_operating', 11, 2)->nullable();
+            $table->decimal('task_budget_gov_investment', 11, 2)->nullable();
+            $table->decimal('task_budget_gov_utility', 11, 2)->nullable();
+            $table->decimal('task_budget_it_operating', 11, 2)->nullable();
+            $table->decimal('task_budget_it_investment', 11, 2)->nullable();
             $table->timestamps();
             $table->softDeletes();
         });
